requirements.txt
Flask
requests
Pillow
python-dotenv
google-generativeai